Modularize poll types, fields, and queues.
