+++ src/tirest/poll/index.ts
@@ -1,0 +1,9 @@
+import { pollPaused } from './paused'
+import { pollPlaying } from './playing'
+
+export { pollPaused, pollPlaying }
+
+export default {
+  paused: pollPaused,
+  playing: pollPlaying,
+}
