+++ src/tirest/poll/paused.ts
@@ -1,0 +1,1 @@
+export function pollPaused(): void {}
